requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.7.2
tweepy==4.14.0
Pillow==10.3.0
python-dotenv==1.0.1
lxml==5.2.2